payroll: deduction gates, rights before the body, first salary from hire, period job fix, test reconciliation
- waive / override / unwaive gate on hr.deductions.create (AD-5, AD-7, AD-8):
  a manager may decide their people's deductions, not only the owner.
- create/decide/stop pay lines check the rights BEFORE judging `kind`
  (AT-11): no 400 that names the fields for someone with no pay rights
  (route_guard_tests was red on it).
- jobs::open_pay_periods read period_start_day as i16 (INT4 → decode panic
  on every tick); i32 now — the money suite caught it.
- Migration 20261001500000: a FIRST salary (old NULL/0) counts from the hire
  date, a raise from the day it is typed (PAY-13); the old staff.rs fixtures
  set the salary after hiring and were paid 8 days of 30.
- tests/staff.rs: paid = every payslip marked paid (no `status: paid` by
  hand), reopen before regenerating, a reason on reopen; tests/dawam_scope.rs:
  the retired legacy pairs replaced by POST /staff/payroll/periods and
  PATCH /staff/advances/{id}/review.

+++ src/staff/jobs.rs
@@ -136,13 +136,15 @@ pub async fn tracking_went_quiet(pool: &PgPool) -> Result<(), crate::errors::App
 /// dashboard read the same row.
 #[doc(hidden)]
 pub async fn open_pay_periods(pool: &PgPool) -> Result<(), crate::errors::AppError> {
-    let orgs: Vec<(Uuid, i16, Option<String>)> = sqlx::query_as(&format!(
+    let orgs: Vec<(Uuid, i32, Option<String>)> = sqlx::query_as(&format!(
         "SELECT o.id, COALESCE(s.period_start_day, 26),                 (SELECT b.timezone::text FROM branches b WHERE b.org_id = o.id AND b.deleted_at IS NULL                   ORDER BY b.created_at LIMIT 1)            FROM organizations o            LEFT JOIN attendance_settings s ON s.org_id = o.id AND s.branch_id IS NULL           WHERE {LIVE_ORG}"
     ))
     .fetch_all(pool)
     .await?;
     for (org_id, start_day, tz) in orgs {
-        let today = crate::staff::attendance::today_in(pool, tz.as_deref().unwrap_or("Africa/Cairo")).await?;
+        let today =
+            crate::staff::attendance::today_in(pool, tz.as_deref().unwrap_or("Africa/Cairo"))
+                .await?;
         crate::staff::dawam::pay::ensure_period_for(pool, org_id, today, start_day.max(1) as u32)
             .await?;
     }
